// src/model/config/PostConverterFilteringOptions.ts
export interface PostConverterFilteringOptions {
  urlsThatEndWithDotJpg: boolean;
  urlsThatEndWithDotJpeg: boolean;
  urlsThatEndWithDotPng: boolean;
  urlsThatEndWithDotGif: boolean;
  urlsInGiphyDomain: boolean;
  urlsInImgurDomain: boolean;
  urlsInRedGifsDomain: boolean;
  redditGalleries: boolean;
}